arm: msm: clock-8226: Add more qcrypto clocks
On 8226, qcrypto can be located at either fd404000 or fd400000.
I've discovered that on my devices (jsr d10f) the instance that
is being used is the one at fd400000 and it couldn't find the
required clocks to initialize.

So, add clocks for that instance for it to be initialized correctly.

Change-Id: I1e2bc9ee279ba62a149e8bf251d38bea4b85732d