block,elevator: use new hashtable implementation
Switch elevator to use the new hashtable implementation. This reduces the
amount of generic unrelated code in the elevator.

This also removes the dymanic allocation of the hash table. The size of the table is
constant so there's no point in paying the price of an extra dereference when accessing
it.

This patch depends on ("hashtable: introduce a small and naive
hashtable") which was merged in v3.6.

Change-Id: I7a4efa2de1fc8a44b76a0d1966910558d9f93eab
Signed-off-by: Sasha Levin <sasha.levin@oracle.com>
Signed-off-by: Jens Axboe <axboe@kernel.dk>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: tarun93 <tarunmyid@gmail.com>
Signed-off-by: hemantbeast <hemantbeast@gmail.com>